refactor: Movie 관련 모델 분리
