+++ The_Oscars/Managers/NetworkManager.swift
@@ -1,0 +1,8 @@
+//
+//  NetworkManager.swift
+//  The_Oscars
+//
+//  Created by eden on 12/17/24.
+//
+
+import Foundation
