+++ The_Oscars/Models/DatesModel.swift
@@ -1,0 +1,14 @@
+//
+//  DatesModel.swift
+//  The_Oscars
+//
+//  Created by eden on 12/17/24.
+//
+
+import Foundation
+
+// MARK: - 날짜 범위 모델
+struct Dates: Codable {
+    let maximum: String
+    let minimum: String
+}
